Add new testing-related images and update testing documentation
Added new images related to project execution and testing workflows, while removing outdated version-related documentation. Simplified and reorganized testing documentation to enhance usability and clarity, focusing on Gradle and testRunner execution processes.

Refactor codebase to adopt event-driven, non-blocking architecture

- transition to a reactive and event-driven architecture.
- Conversion of synchronous MongoDB operations to reactive
-- ensures more efficient and non-blocking data query handling.
- Integration of RabbitMQ and Kafka enables event-based communication and scalability
 - migrated data models from 'record' to classes with Lombok annotations to provide a consistent structure and reduce boilerplate code.

+++ recommendation-service/src/main/java/org/emangini/servolution/core/recommendation/services/RecommendationServiceImpl.java
@@ -11,8 +11,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
-import java.util.List;
+import static java.util.logging.Level.FINE;
 
 @RestController
 @Slf4j
@@ -33,42 +35,59 @@ public class RecommendationServiceImpl implements RecommendationService {
     }
 
     @Override
-    public Recommendation createRecommendation(Recommendation body) {
+    public Mono<Recommendation> createRecommendation(Recommendation body) {
 
-        try {
-            RecommendationEntity entity = mapper.apiToEntity(body);
-            RecommendationEntity savedEntity = repository.save(entity);
-
-            log.debug("createRecommendation: entity created with productId/recommendationId: {}/{}",
-                    body.getProductId(), body.getRecommendationId());
-            return mapper.entityToApi(savedEntity);
-        } catch (DuplicateKeyException e) {
-            throw new InvalidInputException(
-                    "Duplicate key, Product Id: " + body.getProductId()
-                            + ", Recommendation Id: " + body.getRecommendationId()
-            );
+        if (body.getProductId() < 1) {
+            throw new InvalidInputException("Invalid productId: " + body.getProductId());
         }
+
+        RecommendationEntity entity = mapper.apiToEntity(body);
+        return repository.save(entity)
+                .log(log.getName(), FINE)
+                .onErrorMap(
+                        DuplicateKeyException.class,
+                        e -> new InvalidInputException(
+                                "Duplicate key, Product Id: " + body.getProductId()
+                                + ", Recommendation Id: " + body.getRecommendationId())
+                )
+                .map(mapper::entityToApi);
+
     }
 
     @Override
-    public List<Recommendation> getRecommendations(int productId) {
+    public Flux<Recommendation> getRecommendations(int productId) {
 
         if (productId < 1) {
             throw new InvalidInputException("Invalid productId: " + productId);
         }
 
-        List<RecommendationEntity> entities = repository.findByProductId(productId);
-        List<Recommendation> recommendations = mapper.entityListToApiList(entities);
-        recommendations.forEach(recommendation ->
-                recommendation.setServiceAddress(serviceUtil.getServiceAddress()));
+        log.info("calling getRecommendations for product with id={}", productId);
 
-        log.debug("getRecommendations: response size: {}", recommendations.size());
-        return recommendations;
+        return repository.findByProductId(productId)
+                .log(log.getName(), FINE)
+                .map(mapper::entityToApi)
+                .map(this::setServiceAddress);
     }
 
     @Override
-    public void deleteRecommendations(int productId) {
+    public Mono<Void> deleteRecommendations(int productId) {
+
+        if (productId < 1) {
+            throw new InvalidInputException("Invalid productId: " + productId);
+        }
+
         log.debug("deleteRecommendations: attempts to delete an entity with productId: {}", productId);
-        repository.deleteAll(repository.findByProductId(productId));
+
+
+        return repository.deleteAll(repository.findByProductId(productId));
+    }
+
+
+    /*
+    Helper Methods
+ */
+    private Recommendation setServiceAddress(Recommendation recommendation) {
+        recommendation.setServiceAddress(serviceUtil.getServiceAddress());
+        return recommendation;
     }
 }
